Replace SQLNCLI11 library with msoledbsql 18.1.0.
Note: For consistency sake I used an older SDK version, because that's the driver version we install on dev machines. However, it is possible to use a newer SDK while only having an older version of the driver installed. This is an option if we want to update the SDK in the future.
More information about versions can be found here: https://docs.microsoft.com/en-us/sql/connect/oledb/release-notes-for-oledb-driver-for-sql-server

https://jira.ccpgames.com/browse/EO-25233

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/db/": change = 1995130]