Add test project, add a way to override db, make necessary refactors and fix type bugs
A new db test project to verify python3 porting.
Changes to original db project have been kept to a minimum.

Tests are made possible by swapping out OurAccessor for a MockAccessor

Calls to MockAccessor are handled without use of msoledb

Modifications made to db project:
* NSession.h now has a virtual function converter()
This allows us to override this function when creating a MockNSession.
The converter is used for converting DBTYPE_FILETIME to other db date formats
* SQLCommand::SetPyParam visibility moved from private to protected to allow access
* ACCESSOR Macro is used to switch OurAccessor with MockAccessor. Macro route take
rather than implementing purely through templates to keep db code changes to minimum.

Some classes have been refactored by pulling out to separate files.

Done primarily so specific classes can be imported separately.

Downside is that changes are harder to diff.

Pulled out files original location stated in comment at the top of affected file headers

Two unit test files are available:

Tests all the types, entry and retrieval

There is a single method exposed to python 'TestParameter'

TestParmeter takes a db type id and a value to test

The function exercises the preparation of data to be stored for the type

It then exercises the db retrieval processing of that data for the type

The return is then a blue.DBRow with a single element of the retrieved data

Tests check this against expected value

NOTE: This requires a localdb and environment variables set to correctly construct
a connection string.

It can test:
* schema retrieval
* Get/Set session settings
* Execution of a proc

All are currently skipped so TC passes

* DBTYPE_I4 - could result in non graceful python error
* DBTYPE_UI4 - could get stored in container that is too small

Fixed in, in SQLCommand::SetPyParamInt:

```cpp
long long l = PyLong_AsLong( value ); -> long long l = PyLong_AsLongLong( value );
```

Additionally in Row::Init the DBTYPE_UI4 has been moved to unsigned container, though still broken in blue (more info in Broken Types section).

paramLen in SQLCommand::SetPyParam was set to the size of the db types length rather than the length of the input data

Only effects mTotalBytesSentParam which is an exposed python attribute

3 types are still broken for conversation

1. DBTYPE_UI4 is retrieved in PyRowSet.cpp in blue as PyLong_FromLong, this should be PyLong_FromUnsignedLong

2. DBTYPE_CY should allow for 4 decimal places but currently db module only supports setting 2, assuming 2db functionality will persist but left failing for conversation.

3. DBTYPE_IUNKNOWN was broken in PythonBuff, this bit is fixed but more issues remain. left for now as this has been broken since 2005.

Noteworthy: DBTYPE_FILETIME isn't really broken, but when the value is read
in db cextension it is a signed __int64, and when it is retrieved it is unsigned. Tests pass on limits of date info so no big deal, just interesting.

These are known ways the code currently now deviates from python2

Added a pybool type check on input in SQLCommand::SetPyParam, this previously could accept non bool types without error, so could be a compatibility issue. Descriptive error presented if bool check fails.

Byte objects now must be Contiguous. This is checked in code and will complain if not the case

As required by tests there are now more places where type checking is added.