Make DbUnitTests runnable through CTest
When running the tests in test_db.py, all the tests would fail with this cryptic error message:
`SystemError: <built-in function __import__> returned a result with an error set`

This import would fail because a Python error has been set already before doing the import. The error happens when executing an IOWorker request on the main tasklet, thus blocking it. Therefore, change the way tests are run so that they are run using each test files `__main__` function. Set up test_db.py to run it's tests in a non-main tasklet.

PLAT-5236 (Investigate carbon-db's usage of old StacklessIO code)