Fix unit test failing on session count
`testGetSessionStatus` would fail because it expected exactly one session to exist, but it would find two. There is a setting that determines the number of minimum free sessions, and that is by default set to two. Therefore, change the tests to expect the minimum number of free sessions to exist.

PLAT-5236 (Investigate carbon-db's usage of old StacklessIO code)