Speed up IOWorker requests a bit
Complete the request by sending to the channel at the end of the worker thread. Not providing a callback to be called on the loop thread after the work has been completed, since that takes too much time.

Rely on the channel being cleaned up in order to unblock any cancelled requests.